% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_latlon.R
\name{geo_latlon}
\alias{geo_latlon}
\title{Address to the longitude and latitude}
\usage{
geo_latlon(data, add_var = "address", api_key = "")
}
\arguments{
\item{data:}{data frame which must contain the address}

\item{add.var:}{which column or variable contains the address information}

\item{api_key:}{your key for the baidumap api seeing: \url{http://lbsyun.baidu.com/index.php?title=webapi/guide/webservice-geocoding}}
}
\description{
convert the detailed address into the longitude and latitude based on the Baidumap api
}
\examples{
\dontrun{
 geo_latlon(wuhan.sem,add_var = "add",api_key = "sksksksksksk")
}
}
