% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latlon_china.R
\name{get_latlon_china}
\alias{get_latlon_china}
\title{transform the address information into the longitude and latitude}
\usage{
get_latlon_china(data, add_var = "address", api_key = "")
}
\arguments{
\item{data}{data frame, contains the address information}

\item{add_var}{character, variable name in the data, which represents the address information}

\item{api_key}{character,  baidumap api key, seeing: \url{http://lbsyun.baidu.com/index.php?title=webapi/guide/webservice-geocoding}}
}
\value{
two clomuns (lon and lat) was added into the origin data.frame
}
\description{
Based on the Baidumap api, get_latlon_china function coverts the detailed address into the longitude and latitude
}
\examples{
\dontrun{
 get_latlon_china(wuhan.sem,add_var = "add",api_key = "sksksksksksk")
}
}
\author{
Bing Zhang, \url{https://github.com/Spatial-R/EnvExpInd}
}
