% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refrence_data_idw.R
\name{get_refrence_data_idw}
\alias{get_refrence_data_idw}
\title{estimate the environmental exposure using the inverse distance weighting method}
\usage{
get_refrence_data_idw(individual_data, individual_id, exposure_date,
  individual_lat, individual_lon, pollutant_data, pollutant_date = "date",
  pollutant_site_lat, pollutant_site_lon, pollutant_name = c("pm10", "so2"),
  estimate_interval = c(0:30))
}
\arguments{
\item{individual_data}{data.frame includes the refrence id, individual_id and exposure_date}

\item{individual_id}{character, represents the unique id for each individual}

\item{exposure_date}{character, represents the date information}

\item{pollutant_data}{data.frame which indludes the pollutant and site informatin}

\item{pollutant_date}{character,varibale name represents the date infromation for the air pollutant dataset}

\item{pollutant_name}{character, ollutant name in the pollutant_data need to be estimated}

\item{estimate_interval}{continue numeric vector, represents the time-interval to estimate}

\item{pollutant_site}{character,varibale name which includes the site infromation}
}
\value{
aa
}
\description{
used the envronment concentration in the individual location as the reference point to
estimate the environmental exposure. The envionmental concentration at the refrence point was calculated
based on the inverse distance weighting method.
}
\examples{
get_refrence_data_idw(
individual_data = individual_data_tem,
individual_id = "id",
exposure_date ="date",
individual_lat ="lat",
individual_lon ="lon",
pollutant_data = pollutant_data_tem_idw,
pollutant_date = "date",
pollutant_site_lat = "lat",
pollutant_site_lon = "lon",
pollutant_name = c("PM10","PM2.5"),
estimate_interval = c(0:30)
)
}
\author{
Bing Zhang, \url{www.spatial-r.com}
}
