% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuhan_data.R
\docType{data}
\name{individual_data}
\alias{individual_data}
\title{The detailed information for each individual.}
\format{
A data frame with 21 rows and 3 variables:
\describe{
  \item{id}{id number for each individual}
  \item{date}{the monitoring time point}
  \item{lat}{the latitude for each individual}
  \item{lon}{the longtitude for each individual}
  ...
}
}
\usage{
individual_data
}
\description{
A dataset containing the detailed information for each individual
}
\keyword{datasets}
