% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuhan_data.R
\docType{data}
\name{pollutant_data}
\alias{pollutant_data}
\title{The concentration of air pollutant at each time point.}
\format{
A data frame with 11090 rows and 8 variables:
\describe{
  \item{date}{the monitoring time point}
  \item{site.name}{the names of the monitoring site}
  \item{SO2}{the concentration of SO2}
  \item{NO2}{the concentration of NO2}
  \item{PM10}{the concentration of PM10}
  \item{CO}{the concentration of CO}
  \item{O3}{the concentration of O3}
  \item{PM2.5}{the concentration of PM2.5}
  ...
}
}
\usage{
pollutant_data
}
\description{
A dataset containing the concentration of air pollutant at each time point
}
\keyword{datasets}
