% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuhan_data.R
\docType{data}
\name{site_data}
\alias{site_data}
\title{Monitoring sites.}
\format{
A data frame with 10 rows and 2 variables:
\describe{
  \item{site}{the name of monitoring sites}
  \item{lat}{the latitude for each monitoring site}
  \item{lon}{the longtitude for each monitoring site}
  ...
}
}
\usage{
site_data
}
\description{
A dataset containing the information of the monitoring sites
}
\keyword{datasets}
