% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_estimate_idw.R
\name{exposure_estimate_idw}
\alias{exposure_estimate_idw}
\title{estimate the pollutant exposure using the inverse distance weighting method}
\usage{
exposure_estimate_idw(
  individual_data,
  individual_id,
  exposure_date,
  individual_lat,
  individual_lon,
  pollutant_data,
  pollutant_date = "date",
  pollutant_site_lat,
  pollutant_site_lon,
  pollutant_name = c("pm10", "so2"),
  estimate_interval = c(0:30)
)
}
\arguments{
\item{individual_data}{data.frame, contains the refrence id, individual_id and exposure_date}

\item{individual_id}{character,varibale name in individual_data, represents the unique id for each individual}

\item{exposure_date}{character, varibale name in individual_data, which represents the start date to estimate the environment exposure}

\item{individual_lat}{character, varibale name in individual_data, represents the latitude information of each idividual}

\item{individual_lon}{character, varibale name in individual_data, represents the longtitude information of each idividual}

\item{pollutant_data}{data.frame, contains the pollutant and site informatin. One column represents the site information and other columns represent the concentration of pollutants}

\item{pollutant_date}{character, varibale name represents the date infromation for the air pollutant dataset}

\item{pollutant_site_lat}{character, varibale name in pollutant_data, includes the latitude information of each monitoring site}

\item{pollutant_site_lon}{character, varibale name in pollutant_data, includes the longtitude information of each monitoring site}

\item{pollutant_name}{vactor, pollutant name in the pollutant_data, which represent the name of the target pollutants to be estimated}

\item{estimate_interval}{continue numeric vector, the estimation period, for example: 0:30, for each individual we estimate the environment exposure ranging from the exposure_date to exposure_date + 30 days}
}
\value{
a list. every dataframe in the list was with the first column representing the individual id, the remaining columns represent the exposure estimation
in different days.
}
\description{
used the pollutant concentration in the individual location as the reference point to
estimate the environmental exposure. The pollutant concentration at the refrence point was calculated
based on the inverse distance weighting method.
}
\examples{
\dontrun{
library(EnvExpInd)
data(envind)
exposure_estimate_idw(
individual_data = individual_data_tem,
individual_id = "id",
exposure_date ="date",
individual_lat ="lat",
individual_lon ="lon",
pollutant_data = pollutant_data_tem_idw,
pollutant_date = "date",
pollutant_site_lat = "lat",
pollutant_site_lon = "lon",
pollutant_name = c("PM10","PM2.5"),
estimate_interval = c(0:30))
}
}
\author{
Bing Zhang, \url{https://github.com/Spatial-R/EnvExpInd}
}
