% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expoure_estimate_simple.R
\name{expoure_estimate_simple}
\alias{expoure_estimate_simple}
\title{estimate the environmental exposure using the simple method}
\usage{
expoure_estimate_simple(individual_data, individual_id, refrence_id,
  exposure_date, pollutant_data, pollutant_site = "site",
  pollutant_date = "date", pollutant_name = c("pm10", "so2"),
  estimate_interval)
}
\arguments{
\item{individual_data}{the dataset which inludes the refrence id, indi_id and exposure_date}

\item{individual_id}{varibale which includes the id for each individual, fox example: 1,2,3,4}

\item{refrence_id}{varibale which tell you the target site id}

\item{exposure_date}{varibale which has the date information}

\item{pollutant_data}{the dataset which indludes the pollutant and site informatin}

\item{pollutant_site}{varibale name which includes the site infromation}

\item{pollutant_date}{varibale name which includes the date infromation for pollutant}

\item{pollutant_name}{which pollutant need to be estimated}

\item{estimate_interval}{how long do you need to measure, for example: 0:30}
}
\value{
aa
}
\description{
using the nearest surveillance site as the refrence site to estimate the environment exposure.
}
\examples{
expoure_estimate_simple(
    individual_data = individual_data_tem,
   individual_id = "id",
   refrence_id = "refrence_id",
   exposure_date = "date",
   pollutant_data = pollutant_data_tem,
   pollutant_site = "site.name",
   pollutant_date = "date",
   pollutant_name = c("PM10","PM2.5"),
   estimate_interval = c(0:30))
}
\author{
Bing Zhang, \url{www.spatial-r.com}
}
