% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refrence_id_simple.R
\name{get_refrence_id_simple}
\alias{get_refrence_id_simple}
\title{get the refrence site for every sample}
\usage{
get_refrence_id_simple(
  individual_data,
  individual_lat,
  individual_lon,
  individual_id,
  site_data,
  site_lat,
  site_lon,
  site_id
)
}
\arguments{
\item{individual_data}{data.frame from the geo_latlon_china function or geocode function in ggmap}

\item{individual_lat}{character, varibale name in individual_data, includes the latitude information of each idividual}

\item{individual_lon}{character, varibale name in individual_data, includes the longtitude information of each idividual}

\item{individual_id}{character, varibale name in individual_data, includes the unique id for each individual}

\item{site_data}{data.frame from the geo_latlon function}

\item{site_lat}{character varibale includes the latitude value of the site}

\item{site_lon}{character varibale includes the longtitude value of the site}

\item{site_id}{character varibale includes the id for each site}
}
\value{
refrence_id for each individual
}
\description{
get the refrence site for every sample
}
\examples{
\dontrun{
get_refrence_id_simple(
     individual_data = individual_data_tem,
     individual_lat = "lat",
     individual_lon = "lon",
     individual_id = "id",
     site_data = site_data,
     site_lon = "lon",
     site_id = "site")
}
}
\author{
Bing Zhang, \url{https://github.com/Spatial-R/EnvExpInd}
}
